+++ crates/mako/src/config/output.rs
@@ -5,6 +5,7 @@ use clap::ValueEnum;
 use serde::{Deserialize, Serialize};
 use swc_core::ecma::ast::EsVersion;
 
+use super::Umd;
 use crate::create_deserialize_fn;
 use crate::utils::get_app_info;
 
@@ -12,6 +13,7 @@ use crate::utils::get_app_info;
 #[serde(rename_all = "camelCase")]
 pub struct OutputConfig {
     pub path: PathBuf,
+    pub filename: Option<String>,
     pub mode: OutputMode,
     pub es_version: EsVersion,
     pub meta: bool,
@@ -49,8 +51,11 @@ impl fmt::Display for CrossOriginLoading {
     }
 }
 
-pub fn get_default_chunk_loading_global(umd: Option<String>, root: &Path) -> String {
-    let unique_name = umd.unwrap_or_else(|| get_app_info(root).0.unwrap_or("global".to_string()));
+pub fn get_default_chunk_loading_global(umd: Option<Umd>, root: &Path) -> String {
+    let unique_name = umd.map_or_else(
+        || get_app_info(root).0.unwrap_or("global".to_string()),
+        |umd| umd.name.clone(),
+    );
 
     format!("makoChunk_{}", unique_name)
 }