+++ crates/mako/src/module_graph.rs
@@ -2,6 +2,7 @@ use std::collections::{HashMap, HashSet};
 use std::fmt;
 
 use fixedbitset::FixedBitSet;
+use indexmap::IndexSet;
 use petgraph::graph::{DefaultIx, NodeIndex};
 use petgraph::prelude::{Dfs, EdgeRef};
 use petgraph::stable_graph::{StableDiGraph, WalkNeighbors};
@@ -15,7 +16,7 @@ use crate::module::{Dependencies, Dependency, Module, ModuleId};
 pub struct ModuleGraph {
     pub id_index_map: HashMap<ModuleId, NodeIndex<DefaultIx>>,
     pub graph: StableDiGraph<Module, Dependencies>,
-    entries: HashSet<ModuleId>,
+    entries: IndexSet<ModuleId>,
 }
 
 impl ModuleGraph {
@@ -23,7 +24,7 @@ impl ModuleGraph {
         Self {
             id_index_map: HashMap::new(),
             graph: StableDiGraph::new(),
-            entries: HashSet::new(),
+            entries: IndexSet::new(),
         }
     }