+++ crates/mako/src/plugins/bundless_compiler.rs
@@ -131,7 +131,7 @@ impl BundlessCompiler {
 
         self.context
             .plugin_driver
-            .before_write_fs(&to, content.as_ref())
+            .before_write_fs(&to, content.as_ref(), &self.context)
             .unwrap();
 
         if !self.context.config.output.skip_write {