+++ e2e/fixtures/css.css-merge-multi-entries/src/b.ts
@@ -1,0 +1,5 @@
+import "./common_2.css";
+import "./common_1.css";
+import "./common_3.css";
+import "./b.css";
+console.log('b');