+++ examples/import-resources/import-js-wasm/index.js
@@ -1,0 +1,5 @@
+import * as wasm from './index_bg.wasm';
+export * from './index_bg.js';
+import { __wbg_set_wasm } from './index_bg.js';
+__wbg_set_wasm(wasm);
+wasm.__wbindgen_start();