+++ examples/import-resources/minus-wasm-pack/index.d.ts
@@ -1,0 +1,15 @@
+/* tslint:disable */
+/* eslint-disable */
+/**
+ */
+export function greet(): void;
+/**
+ * @param {string} name
+ */
+export function greet2(name: string): void;
+/**
+ * @param {number} a
+ * @param {number} b
+ * @returns {number}
+ */
+export function minus(a: number, b: number): number;